pyproject.toml
[tool.poetry]
name = "dcf"
version = "0.1.0"
description = ""
authors = [
    "Jeppe Klitgaard <jk782@cam.ac.uk>",
    "Tom Ben-Ami <tab63@cam.ac.uk>"
]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
opencv-python = "^4.5.4"
Pillow = "^8.4.0"
pandas = "^1.3.4"
matplotlib = "^3.4.3"
cddm = "^0.3.0"
pyffmpeg = "^2.0.5"

[tool.poetry.dev-dependencies]
ipykernel = "^6.5.0"
json5 = "^0.9.6"
pyjson5 = "^1.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
